Update
-Generator:
Added cable to generator.
Handle should be grabbable and when released should go back.
Handle should have limit to how far it can be stretched (SpringJoint)

Pipes-
Added pipes model.
Added sound emitter and set distance so that you must be close to hear it.
Made pattern values public so that it can be accessed and then used for which candles are supposed to be lit.

Doors-
Attempted adding door functionality with diff techniques, should be grabbable but we might need to work on reducing bugginess.

TL;DR - Doors, generator, pipes. A lotta stuff that I can't test without a working headset.